Replace laravel/framework with individual packages (#304)
Rather than depend on the entire laravel framework, instead
composer.json now gives each Illuminate component needed.

This makes installation in Lumen much easier, as pulling in the
laravel framework causes issues, however all the required illuminate
components are already there as dependencies of the Lumen framework.

Dev requirements still depend on laravel/framework. This does not cause
any issues when the package is required in a Lumen project as these do
not get pulled in.

The only source code change is to use Carbon::now() rather than the
Laravel now() helper function - to be Laravel / Lumen agnostic.

The provider class can be used as is in Lumen.